complete step 1 - basic app structure with two GET endpoints public and protected
